error free for now.  Cleaned up the time taken in seconds, switched to a long to be safe

+++ backend/src/main/java/com/example/Unit_2_Project/dto/QuizAttemptSummaryDTO.java
@@ -11,5 +11,5 @@ public class QuizAttemptSummaryDTO {
 
     private int score; // Final score
 
-    private int timeTakenInSeconds; // Time taken to complete quiz
+    private long timeTakenInSeconds; // Time taken to complete quiz
 }
